+++ src/main/java/org/killbill/billing/catalog/api/TieredBlockPriceOverride.java
@@ -18,9 +18,11 @@
 
 package org.killbill.billing.catalog.api;
 
+import java.math.BigDecimal;
+
 public interface TieredBlockPriceOverride extends BlockPriceOverride {
     /**
      * @return the max used usage to consider this tiered block.
      */
-    public Double getMax();
+    public BigDecimal getMax();
 }
